+++ python/lsst/ts/weatherforecast/csc.py
@@ -23,6 +23,7 @@ __all__ = ["WeatherForecastCSC", "execute_csc"]
 
 import asyncio
 import datetime
+import math
 import os
 
 import aiohttp
@@ -63,6 +64,10 @@ class WeatherForecastCSC(salobj.ConfigurableCsc):
     simulation_mode : `int`
         Whether the CSC is in simulation mode.
 
+        * 0 - real mode
+        * 1 - simulated data
+        * 2 - simulated missing data
+
     Attributes
     ----------
     telemetry_task : `asyncio.Future`
@@ -77,6 +82,15 @@ class WeatherForecastCSC(salobj.ConfigurableCsc):
         The wait time for retrying if the API call fails.
     api_key : `str`
         The stored API key for Meteoblue received from an environment variable.
+
+    Notes
+    -----
+
+    Simulation mode has three values.
+
+    * 0 - real data mode.
+    * 1 - Simulated data mode.
+    * 2 - Missing data mode.
     """
 
     valid_simulation_modes = [0, 1, 2]
@@ -119,22 +133,52 @@ class WeatherForecastCSC(salobj.ConfigurableCsc):
         """Convert string to datetime object and then convert to unix
         timestamp.
 
+        This is used to convert the string that MeteoBlue returns for time
+        into a timestamp that can be published over DDS.
+
         Parameters
         ----------
         timestamp : `str`
             The time to convert.
+
+        Returns
+        -------
+        timestamp : `list` of `int`
+            An array of timestamps converted from the date string.
         """
         timestamp = datetime.datetime.strptime(timestamp, "%Y-%m-%d %H:%M").timestamp()
         return timestamp
 
-    def pad_data(self, lst, match):
-        self.log.debug("Padding values according to type.")
+    def fix_data_length(self, lst, match):
+        """Pad the data with default values.
+
+        MeteoBlue's API sometimes returns inconsistent count of values.
+        This method is an attempt to sanitize the data.
+
+        Parameters
+        ----------
+        lst : `list`
+            List of values to pad out.
+        match : `int`
+            The count to match the length of the list to.
+            Hourly trend is 382 and daily trend is 15.
+
+        Returns
+        -------
+        lst : `list`
+            The corrected count list of values.
+        """
+        self.log.info("Attempting to pad data in order to continue.")
+        if len(lst) > match:
+            self.log.info("Ignoring excess values.")
+            lst = lst[:match]
+            return lst
         if isinstance(lst[0], int):
-            lst.extend([0] * match)
+            lst.extend([-1] * (match - len(lst)))
         elif isinstance(lst[0], float):
-            lst.extend([0.0] * match)
+            lst.extend([math.nan] * (match - len(lst)))
         elif isinstance(lst[0], str):
-            lst.extend(["1970-01-01 00:00"] * match)
+            lst.extend(["1970-01-01 00:00"] * (match - len(lst)))
         else:
             raise RuntimeError(f"Unable to pad type {type(lst[0])}.")
         lst = lst[:match]
@@ -184,23 +228,24 @@ class WeatherForecastCSC(salobj.ConfigurableCsc):
                     modelrunUpdatetime=modelrun_updatetime_utc,
                 )
                 trend_hour_fld = response["trend_1h"]
-                for name, values in trend_hour_fld.items():
-                    if len(values) == COUNT_HOURLY:
-                        pass
-                    else:
-                        self.log.error(
-                            f"Count of {name} = {len(values)}, should be {COUNT_HOURLY}."
-                        )
-                        self.log.info("Attempting to pad data in order to continue.")
-                        trend_hour_fld[name] = self.pad_data(values, COUNT_HOURLY)
-                timestamps = trend_hour_fld["time"]
-                converted_timestamps = [
-                    self.convert_time(timestamp) for timestamp in timestamps
-                ]
                 # check for None in extraTerrestrialRadiationBackwards
                 trend_hour_fld["extraterrestrialradiation_backwards"] = [
                     0.0 if value is None else value
                     for value in trend_hour_fld["extraterrestrialradiation_backwards"]
+                ]
+                for name, values in trend_hour_fld.items():
+                    if len(values) == COUNT_HOURLY:
+                        pass
+                    else:
+                        self.log.warning(
+                            f"Count of {name} = {len(values)}, setting it to {COUNT_HOURLY}."
+                        )
+                        trend_hour_fld[name] = self.fix_data_length(
+                            values, COUNT_HOURLY
+                        )
+                timestamps = trend_hour_fld["time"]
+                converted_timestamps = [
+                    self.convert_time(timestamp) for timestamp in timestamps
                 ]
                 await self.tel_hourlyTrend.set_write(
                     timestamp=converted_timestamps,
@@ -240,21 +285,20 @@ class WeatherForecastCSC(salobj.ConfigurableCsc):
                     ],
                 )
                 trend_daily_fld = response["trend_day"]
-                timestamps = trend_daily_fld["time"]
-                converted_timestamps = [
-                    self.convert_time(timestamp) for timestamp in timestamps
-                ]
                 for name, values in trend_daily_fld.items():
                     if len(values) == COUNT_DAILY:
                         pass
                     else:
                         self.log.error(
-                            f"Count of {name} = {len(values)}, should be {COUNT_DAILY}"
+                            f"Count of {name} = {len(values)}, setting it to {COUNT_DAILY}."
                         )
-                        trend_daily_fld[name] = self.pad_data(values, COUNT_DAILY)
-                        raise RuntimeError(
-                            f"Count of {name} = {len(values)}, should be {COUNT_DAILY}."
+                        trend_daily_fld[name] = self.fix_data_length(
+                            values, COUNT_DAILY
                         )
+                timestamps = trend_daily_fld["time"]
+                converted_timestamps = [
+                    self.convert_time(timestamp) for timestamp in timestamps
+                ]
                 await self.tel_dailyTrend.set_write(
                     timestamp=converted_timestamps,
                     pictocode=trend_daily_fld["pictocode"],