+++ python/lsst/ts/weatherforecast/mock_server.py
@@ -46,6 +46,9 @@ class MockServer:
         The webapp runner.
     site : `None`
         The webapp site object.
+    bad_request_counter : `int`
+        Meant to count the number of bad requests to send before returning
+        a good response.
     """
 
     def __init__(
@@ -55,7 +58,7 @@ class MockServer:
         self.runner = None
         self.site = None
         test_file = pathlib.Path(data)
-        self.retries = 0
+        self.bad_request_counter = 0
         self.log = logging.getLogger(__name__)
         with open(test_file) as f:
             self.response = json.load(f)
@@ -111,8 +114,8 @@ class MockServer:
             The canned json response.
             See the test file in the data directory for the format.
         """
-        while self.retries <= 3:
-            self.log.info(f"Inside bad request check. {self.retries=}")
-            self.retries += 1
+        if self.bad_request_counter <= 3:
+            self.log.info(f"Inside bad request check. {self.bad_request_counter=}")
+            self.bad_request_counter += 1
             raise web.HTTPBadRequest(reason="Something went wrong.")
         return web.json_response(self.response)